sprint-2/02-inheritance/demo_inheritance/src/main/java/com/niantic/shapes/Rectangle.java
package com.niantic.shapes;

public class Rectangle
{ //default values
    private double height = 10;
    private double width = 10;

    public double getHeight()
    {
        return height;
    }

    public void setHeight(double height)
    {
        this.height = height;
    }

    public double getWidth()
    {
        return width;
    }

    public void setWidth(double width)
    {
        this.width = width;
    }

    public double getArea()
    {
        return width * height;
    }

    public double getPerimeter()
    {
        return 2 * width + 2 * height;
    }
}
